refactoring serializers
