+++ bangazon/bangazonapi/serializers/__init__.py
@@ -1,0 +1,1 @@
+__all__ = ['customer_serializer', 'order_serializer', 'paymenttype_serializer', 'product_on_order_serializer', 'product_serializer', 'product_types_serializer']
