+++ bangazon/bangazonapi/serializers/order_serializer.py
@@ -1,0 +1,20 @@
+from rest_framework import serializers
+from bangazonapi.models import *
+
+class OrderSerializer(serializers.HyperlinkedModelSerializer):
+    ''' The OrderSerializer class translates the Order models into other formats, in this case JSON by default. that Order table so a database can be created from it.
+
+    Method List:
+    -Meta
+    -create
+    -update
+
+    Argument List:
+    -serializers.HyperlinkedModelSerializer: This argument allows the class to access field types.
+
+    Author: Zoe LeBlanc, Python Ponies
+    '''
+    class Meta:
+        model = order_model.Order
+        fields = '__all__'
+        depth = 3
