+++ bangazon/bangazonapi/serializers/product_types_serializer.py
@@ -1,0 +1,19 @@
+from rest_framework import serializers
+from bangazonapi.models import *
+
+class ProductCategorySerializer(serializers.HyperlinkedModelSerializer):
+    ''' The CategorySerializer class translates the Category models into other formats, in this case JSON by default. that Category table so a database can be created from it.
+
+    Method List:
+    -Meta
+    -create
+    -update
+
+    Argument List:
+    -serializers.HyperlinkedModelSerializer: This argument allows the class to access field types.
+
+    Author: Zoe LeBlanc, Python Ponies
+    '''
+    class Meta:
+        model = product_types_model.ProductTypes
+        fields = '__all__'
